+++ src/utils/check-name.ts
@@ -1,6 +1,6 @@
 import { players } from '@/db/in-memory-db';
 
-export const isUniqueName = (name: string) => {
+export const checkName = (name: string) => {
   for (const [_, player] of players.entries()) {
     if (player.name.trim() === name.trim()) {
       return true;
