+++ src/utils/get-rooms.ts
@@ -5,7 +5,7 @@ export const getRooms = () => {
   const rooms: Room[] = [];
 
   for (const [roomId, game] of games) {
-    if (game.second.ws.index === game.first.ws.index) {
+    if (game.second.ws?.index === game.first.ws?.index) {
       rooms.push({ roomId, roomUsers: [{ name: game.first.ws!.name, index: game.first.ws!.index }] });
     }
   }
