+++ src/utils/init-board.ts
